Move install_packages resource to platform cookbook
Also:
- add :packages property
- add :install action, which installs packages passed via :packages property, or the default set of packages if the property is not set
- move dns setup out of this resource, since it is currently related to slurm cookbook and orthogonal to the base packages
- we currently have build_tools and modules resources which are subsets of base packages; in the future we might want to invest into further split of base packages into clearly defined, cohesive package subsets

Test:
- `bash kitchen.ec2.sh platform-install verify install-packages -c 6`: OK
- `bash kitchen.docker.sh platform-install verify install-packages -c 6`: OK
